il reste follow user

+++ controllers/postController.js
@@ -1,11 +1,12 @@
 const { isEmpty } = require("../lib/allFunctions");
 const { isValidObjectId } = require("mongoose");
+const { io, getReceiverSocketId } = require("../socket");
 
 const UserModel = require("../models/UserModel");
 const PostModel = require("../models/PostModel");
 const fs = require("fs");
 const path = require("path");
-const { io } = require("../socket");
+const NotificationModel = require("../models/NotificationModel");
 
 module.exports.getAllPosts = async (req, res) => {
   try {
@@ -21,17 +22,6 @@ module.exports.getAllPosts = async (req, res) => {
     }
 
     const posts = await PostModel.find().sort({ updatedAt: -1 });
-
-    // posts.forEach(async (p) => {
-    //   let user = await UserModel.findById(p.senderId);
-    //   if (!user.posts.includes(p._id)) {
-    //     user = await UserModel.findByIdAndUpdate(
-    //       p.senderId,
-    //       { $push: { posts: p._id } },
-    //       { new: true }
-    //     );
-    //   }
-    // });
 
     return res.status(200).json({ posts });
   } catch (error) {
@@ -64,6 +54,7 @@ module.exports.getMyPosts = async (req, res) => {
 
 module.exports.getPost = async (req, res) => {
   try {
+    let user = null;
     const params = req.params;
 
     if (isEmpty(params?.id) || !isValidObjectId(params?.id)) {
@@ -72,14 +63,27 @@ module.exports.getPost = async (req, res) => {
       return res.json({ postIdRequired: true });
     }
 
-    const user = await UserModel.findById(params.id);
+    user = await UserModel.findById(params.id);
     if (isEmpty(user)) {
       return res.json({ userNotFound: true });
     }
 
     const post = await PostModel.findById(params.postId);
+    if (isEmpty(post)) {
+      return res.json({ postNotFound: true });
+    }
 
-    return res.status(200).json({ post });
+    if (user.rejectedPost.includes(params.postId)) {
+      user = await UserModel.findByIdAndUpdate(
+        params.id,
+        { $pull: { rejectedPost: params.postId } },
+        { new: true }
+      );
+    }
+
+    return res
+      .status(200)
+      .json({ post, user: { rejectedPost: user.rejectedPost } });
   } catch (error) {
     return res.status(500).json({ error: `${error.message}` });
   }
@@ -89,6 +93,7 @@ module.exports.likePost = async (req, res) => {
   try {
     let user = null;
     let post = null;
+    let notification = null;
     const params = req.params;
 
     if (isEmpty(params?.id) || !isValidObjectId(params?.id)) {
@@ -107,18 +112,64 @@ module.exports.likePost = async (req, res) => {
       return res.json({ postNotFound: true });
     }
 
+    const receiverSocketId = getReceiverSocketId(post.senderId);
+
     if (post.likes.includes(params.id)) {
       post = await PostModel.findByIdAndUpdate(
         params.postId,
         { $pull: { likes: params.id } },
         { new: true }
       );
+
+      if (post.senderId !== params.id) {
+        notification = await NotificationModel.findOne({
+          userId: post.senderId,
+          senderId: params.id,
+          postId: params.postId,
+          liked: true,
+        });
+
+        if (notification) {
+          await NotificationModel.findByIdAndDelete(notification._id);
+          io.to(receiverSocketId).emit("unlikePostNotification", notification);
+        }
+      }
+
+      io.emit("unlikePost", post);
     } else {
       post = await PostModel.findByIdAndUpdate(
         params.postId,
         { $addToSet: { likes: params.id } },
         { new: true }
       );
+
+      if (post.senderId !== params.id) {
+        notification = await NotificationModel.findOne({
+          userId: post.senderId,
+          senderId: params.id,
+          postId: params.postId,
+          liked: true,
+        });
+
+        if (notification) {
+          notification = await NotificationModel.findByIdAndUpdate(
+            notification._id,
+            { $set: { viewed: false } },
+            { new: true }
+          );
+        } else {
+          notification = await NotificationModel.create({
+            userId: post.senderId,
+            senderId: params.id,
+            postId: params.postId,
+            liked: true,
+          });
+        }
+
+        io.to(receiverSocketId).emit("likePostNotification", notification);
+      }
+
+      io.emit("likePost", post);
     }
 
     if (user.rejectedPost.includes(params.postId)) {
@@ -128,8 +179,6 @@ module.exports.likePost = async (req, res) => {
         { new: true }
       );
     }
-
-    io.emit("likePost", post);
 
     return res
       .status(200)
@@ -141,7 +190,8 @@ module.exports.likePost = async (req, res) => {
 
 module.exports.commentPost = async (req, res) => {
   try {
-    let post = {};
+    let post = null;
+    let notification = null;
     const body = req.body;
     const params = req.params;
 
@@ -177,6 +227,52 @@ module.exports.commentPost = async (req, res) => {
       { new: true }
     );
 
+    if (post.senderId !== params.id) {
+      const receiverSocketId = getReceiverSocketId(post.senderId);
+
+      notification = await NotificationModel.findOne({
+        userId: post.senderId,
+        senderId: params.id,
+        postId: params.postId,
+        commented: true,
+      });
+
+      if (notification) {
+        notification = await NotificationModel.findByIdAndUpdate(
+          notification._id,
+          { $set: { viewed: false } },
+          { new: true }
+        );
+      } else {
+        notification = await NotificationModel.findOne({
+          userId: post.senderId,
+          senderId: params.id,
+          postId: params.postId,
+          commented: true,
+          viewed: false,
+        });
+
+        if (notification) {
+          notification = await NotificationModel.findByIdAndUpdate(
+            notification._id,
+            { $set: { viewed: false } },
+            { new: true }
+          );
+        } else {
+          notification = await NotificationModel.create({
+            userId: post.senderId,
+            senderId: params.id,
+            postId: params.postId,
+            commented: true,
+            viewed: false,
+          });
+        }
+      }
+
+      io.to(receiverSocketId).emit("commentPostNotification", notification);
+    }
+    io.emit("commentPost", post);
+
     if (user.rejectedPost.includes(params.postId)) {
       user = await UserModel.findByIdAndUpdate(
         params.id,
@@ -184,8 +280,6 @@ module.exports.commentPost = async (req, res) => {
         { new: true }
       );
     }
-
-    io.emit("commentPost", post);
 
     return res
       .status(200)
@@ -199,6 +293,7 @@ module.exports.deleteComment = async (req, res) => {
   try {
     let user = null;
     let post = null;
+    let notification = null;
     const params = req.params;
 
     if (isEmpty(params?.id) || !isValidObjectId(params?.id)) {
@@ -224,11 +319,40 @@ module.exports.deleteComment = async (req, res) => {
       return res.json({ postNotFound: true });
     }
 
+    const comments = post.comments.filter(
+      (comment) => comment.userId === params.id
+    );
+
     post = await PostModel.findByIdAndUpdate(
       params.postId,
       { $pull: { comments: { _id: params.commentId } } },
       { new: true }
     );
+
+    if (post.senderId !== params.id) {
+      const receiverSocketId = getReceiverSocketId(post.senderId);
+
+      notification = await NotificationModel.findOne({
+        userId: post.senderId,
+        senderId: params.id,
+        postId: params.postId,
+        commented: true,
+      });
+
+      if (notification) {
+        if (!isEmpty(comments) && comments?.length === 1) {
+          notification = await NotificationModel.findByIdAndDelete(
+            notification._id
+          );
+          io.to(receiverSocketId).emit(
+            "deleteCommentPostNotification",
+            notification
+          );
+        }
+      }
+    }
+
+    io.emit("deleteCommentPost", post);
 
     if (user.rejectedPost.includes(params.postId)) {
       user = await UserModel.findByIdAndUpdate(
@@ -237,8 +361,6 @@ module.exports.deleteComment = async (req, res) => {
         { new: true }
       );
     }
-
-    io.emit("deleteCommentPost", post);
 
     return res
       .status(200)
@@ -252,6 +374,7 @@ module.exports.createPost = async (req, res) => {
   try {
     let fileName = null;
     let user = null;
+    let users = null;
     const params = req.params;
     const body = req.body;
 
@@ -298,7 +421,120 @@ module.exports.createPost = async (req, res) => {
       { new: true }
     );
 
-    io.emit("createPost", post);
+    users = await UserModel.find({ _id: { $ne: params.id } });
+    users.forEach(async (item) => {
+      const receiverSocketId = getReceiverSocketId(item._id);
+      const notification = await NotificationModel.create({
+        userId: item._id,
+        senderId: params.id,
+        postId: params.postId,
+        newPost: true,
+        viewed: false,
+      });
+
+      io.to(receiverSocketId).emit("createPost", notification);
+    });
+
+    return res.status(200).json({ post });
+  } catch (error) {
+    return res.status(500).json({ error: `${error.message}` });
+  }
+};
+
+module.exports.editPost = async (req, res) => {
+  try {
+    let fileName = null;
+    let user = null;
+    let post = null;
+    let users = null;
+    const params = req.params;
+    const body = req.body;
+
+    if (isEmpty(params?.id) || !isValidObjectId(params?.id)) {
+      return res.json({ idRequired: true });
+    } else if (isEmpty(params?.postId) || !isValidObjectId(params?.postId)) {
+      return res.json({ postIdRequired: true });
+    }
+
+    user = await UserModel.findById(params.id);
+    if (isEmpty(user)) {
+      return res.json({ userNotFound: true });
+    }
+
+    post = await PostModel.findById(params.postId);
+    if (isEmpty(post)) {
+      return res.json({ postNotFound: true });
+    }
+
+    if (isEmpty(body?.message) && !req.file) {
+      return res.json({ dataRequired: true });
+    }
+
+    let infos = {};
+    infos.senderId = params.id;
+
+    if (!isEmpty(body?.message?.trim())) {
+      infos.message = body.message.trim();
+    }
+
+    if (req.file) {
+      if (
+        req.file.mimetype !== "image/jpg" &&
+        req.file.mimetype !== "image/jpeg" &&
+        req.file.mimetype !== "image/png"
+      ) {
+        return res.json({ invalidFormat: true });
+      }
+      if (!isEmpty(post.image)) {
+        const imagePath = path.join(__dirname, "../uploads/post", post.image);
+        if (fs.existsSync(imagePath)) {
+          fs.unlinkSync(imagePath);
+        }
+      }
+
+      fileName = user._id + "-" + Date.now() + ".jpg";
+      const filePath = path.join(__dirname, "../uploads/post", fileName);
+      fs.writeFileSync(filePath, req.file.buffer);
+
+      infos.image = fileName;
+    }
+
+    post = await PostModel.findByIdAndUpdate(
+      params.postId,
+      { $set: infos },
+      { new: true }
+    );
+
+    users = await UserModel.find({ _id: { $ne: params.id } });
+    users.forEach(async (item) => {
+      const receiverSocketId = getReceiverSocketId(item._id);
+      let notification = await NotificationModel.findOne({
+        userId: item._id,
+        senderId: params.id,
+        postId: params.postId,
+        editPost: true,
+      });
+
+      if (notification) {
+        notification = await NotificationModel.findByIdAndUpdate(
+          notification._id,
+          { $set: { viewed: false } },
+          { new: true }
+        );
+      } else {
+        notification = await NotificationModel.create({
+          userId: item._id,
+          senderId: params.id,
+          postId: params.postId,
+          editPost: true,
+          viewed: false,
+        });
+      }
+
+      io.to(receiverSocketId).emit("editPostNotification", notification);
+    });
+
+    io.emit("editPost", post);
 
     return res.status(200).json({ post });
   } catch (error) {
@@ -369,12 +605,16 @@ module.exports.deletePost = async (req, res) => {
       return res.json({ notYourPost: true });
     }
 
+    if (!isEmpty(post?.image)) {
+      const imagePath = path.join(__dirname, "../uploads/post", post.image);
+      if (fs.existsSync(imagePath)) {
+        fs.unlinkSync(imagePath);
+      }
+    }
+
     post = await PostModel.findByIdAndDelete(params.postId);
 
-    const imagePath = path.join(__dirname, "../uploads/post", post.image);
-    if (fs.existsSync(imagePath)) {
-      fs.unlinkSync(imagePath);
-    }
+    io.emit("deletePost", post);
 
     return res.status(200).json({ post });
   } catch (error) {
